MAINT make it possible to use the system joblib
