+++ sklearn/externals/copy_joblib.sh
@@ -21,4 +21,16 @@ find joblib -name "*.bak" | xargs rm
 # joblib is already tested on its own CI infrastructure upstream.
 rm -r joblib/test
 
+# Make it possible to use the system joblib based on an environment variable
+mv joblib/__init__.py joblib/embedded_joblib_init.py
+NEW_INIT="
+import os
+if int(os.environ.get('SKLEARN_SYSTEM_JOBLIB', '0')):
+    # Try to import joblib from the system for development purpose
+    from joblib import *
+else:
+    from .embedded_joblib_init import *
+"
+echo "$NEW_INIT" > joblib/__init__.py
+
 chmod -x joblib/*.py
